backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
chromadb==0.4.18
openai==1.3.7
python-multipart==0.0.6
GitPython==3.1.40
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
tenacity==8.2.3
sentence-transformers==2.2.2
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.2
tree-sitter-typescript==0.20.3
regex==2023.10.3
tiktoken==0.5.2
